requirements.txt
astral==2.2
python-dotenv==0.20.0
pytz==2022.1
schedule==1.1.0
watchdog==2.1.9
